rename occurence to effective date
