+++ source/Energinet.DataHub.MeteringPoints.Application/Validation/ValidationErrors/EffectiveDateRequiredValidationError.cs
@@ -16,7 +16,7 @@ using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
 namespace Energinet.DataHub.MeteringPoints.Application.Validation.ValidationErrors
 {
-    public class OccurenceRequiredValidationError : ValidationError
+    public class EffectiveDateRequiredValidationError : ValidationError
     {
     }
 }
